requirements.txt
streamlit
pdfplumber
chromadb
google-generativeai
pandas
matplotlib